mlops/APIs/fr-api/pyproject.toml
[project]
name = "fr-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "marshmallow==3.19.0",
    "numpy>=2.2.6",
    "pillow>=12.1.0",
    "pymilvus==2.2.15",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "setuptools>=80.10.1",
    "uvicorn[standard]>=0.40.0",
]
